Refactor Backend Folders/Packages
