--- server/Config/default.json
@@ -1,3 +1,0 @@
-{
-  "mongoURI": "mongodb://127.0.0.1:27017/hackshare?gssapiServiceName=mongodb"
-}
