+++ server/Model/index.js
@@ -1,0 +1,19 @@
+const user = require("./user");
+const task = require("./task");
+const meeting = require("./meeting");
+const learning = require("./learning");
+const expertise = require("./expertise");
+const endorsement = require("./endorsement");
+const challenge = require("./challenge");
+const achievement = require("./achievement");
+
+module.exports = {
+  user,
+  task,
+  meeting,
+  learning,
+  expertise,
+  endorsement,
+  challenge,
+  achievement,
+};
