+++ server/README.md
@@ -4,9 +4,9 @@ This is the Backend folder, all API, Authentication etc are in this folder.
 
 ### Folder Layout
 
-#### Config
+#### Helper
 
-This is a configuration file, with the help of `config` package, this folder has global variables, which can be access everywhere. For example, MongoDB URI.
+This folder contains all the helper, High Order Functions functions.
 
 #### Controller
